refactor: split file manager data api
